why are we using oxyplot instead of sci chart, why not just use one or the other. also why do we have so much code behind.

+++ View/Output/HistogramViewer.xaml.cs
@@ -55,7 +55,7 @@ namespace HEC.FDA.View.Output
 
         }
 
-        private void setUpHistogramPlotModel( int sampleSize, List<HistogramBinVM> bins)
+        private void setUpHistogramPlotModel( Int64 sampleSize, List<HistogramBinVM> bins)
         {
 
             HistogramPlotModel = new OxyPlot.PlotModel();
